adds login features

+++ controllers/movies_router.js
@@ -1,9 +1,10 @@
 const moviesRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 const Movie = require('../models/movie');
 const User = require('../models/user');
 
 moviesRouter.get('/', async (request, response) => {
-  const movies = await Movie.find({});
+  const movies = await Movie.find({}).populate('user');
   response.json(movies.map((movie) => movie.toJSON()));
 });
 
@@ -16,7 +17,11 @@ moviesRouter.get('/:id', async (request, response) => {
 
 moviesRouter.post('/', async (request, response) => {
   const body = request.body;
-  const user = await User.findById(request.body.userId);
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id)
+    return response.status(401).json({ error: 'token missing or invalid' });
+  const user = await User.findById(body.userId);
 
   const movie = new Movie({
     title: body.title,
@@ -33,16 +38,38 @@ moviesRouter.post('/', async (request, response) => {
 });
 
 moviesRouter.delete('/:id', async (request, response) => {
+  const token = request.token;
+  const movie = await Movie.findById(request.params.id);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token) return response.status(401).json({ error: 'token missing' });
+  else if (decodedToken.id !== movie.user.toString())
+    return response.status(401).json({ error: 'unauthorized access.' });
+
   await Movie.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
 
 moviesRouter.put('/:id', async (request, response) => {
-  const movie = request.body;
+  const body = request.body;
+  const token = request.token;
+  const movie = await Movie.findById(request.params.id);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
 
-  const updated = await Movie.findByIdAndUpdate(request.params.id, movie, {
-    new: true,
-  });
+  if (!token) return response.status(401).json({ error: 'token missing' });
+  else if (decodedToken.id !== movie.user.toString())
+    return response.status(401).json({ error: 'unauthorized access.' });
+
+  const updatedMovie = {
+    score: body.score,
+  };
+  const updated = await Movie.findByIdAndUpdate(
+    request.params.id,
+    updatedMovie,
+    {
+      new: true,
+    }
+  );
   response.json(updated.toJSON);
 });
